+++ src/rtp.cc
@@ -167,11 +167,11 @@ void uvgrtp::rtp::fill_header(uint8_t *buffer, bool use_old_ts)
 
     if (use_old_ts) {
         if (timestamp_ == INVALID_TS) {
-            *(uint32_t*)&buffer[4] = htonl((u_long)rtp_ts_);
+            *(uint32_t*)&buffer[4] = htonl(static_cast<uint32_t>(rtp_ts_));
         }
         else {
-            *(uint32_t*)&buffer[4] = htonl((u_long)timestamp_);
-            rtp_ts_ = timestamp_;
+            *(uint32_t*)&buffer[4] = htonl(static_cast<uint32_t>(timestamp_));
+            rtp_ts_ = static_cast<uint32_t>(timestamp_);
 
             if (!custom_ntp_)
             {
@@ -191,12 +191,12 @@ void uvgrtp::rtp::fill_header(uint8_t *buffer, bool use_old_ts)
         rtp_ts_ = rtp_timestamp;
         sampling_ntp_ = uvgrtp::clock::ntp::now();
 
-        *(uint32_t *)&buffer[4] = htonl((u_long)rtp_timestamp);
+        *(uint32_t *)&buffer[4] = htonl(static_cast<uint32_t>(rtp_timestamp));
 
     }
     else {
-        *(uint32_t *)&buffer[4] = htonl((u_long)timestamp_);
-        rtp_ts_ = timestamp_;
+        *(uint32_t *)&buffer[4] = htonl(static_cast<uint32_t>(timestamp_));
+        rtp_ts_ = static_cast<uint32_t>(timestamp_);
 
         if (!custom_ntp_)
         {
